@@ -7,7 +7,7 @@ namespace DataLayer.Repositories.Interfaces;
 public interface IAuthRepository
 {
     Task<IEnumerable<User>> AddMockUsersAsync(int numOfUsers = 10);
-    Task<(JwtToken, JwtToken)> LoginAsync(LoginProp prop);
-    Task<User> RegisterUserAsync(RegisterDTO registerDto);
+    Task<(JwtToken, JwtToken)> LoginAsync(string loginOrEmail, string password);
+    Task<User> RegisterUserAsync(string login, string email, string password);
     Task<(JwtToken, JwtToken)> RefreshToken(string refreshToken);
 }
